fix: errors on last line now work

+++ liamc/src/lexer.h
@@ -72,9 +72,9 @@ std::string get_token_type_string(TokenType type);
 struct Span {
     // lines start from 0 always
     // start which is the starting character of the span start from 0
-    u32 line;
-    u32 start;
-    u32 end;
+    u64 line;
+    u64 start;
+    u64 end;
 };
 
 struct Token {
